backend/utils/db.py
# backend/utils/db.py
from pymongo import MongoClient
from dotenv import load_dotenv
import os

load_dotenv()

client = MongoClient(os.getenv('MONGO_URI'))
db = client[os.getenv('DB_NAME')]

def init_db():
    # Initialize collections if needed
    pass
